[chore] relocate create plan client files
