+++ src/features/plan/base/create/schemas/formSchema.ts
@@ -1,7 +1,8 @@
 import { z } from "zod";
+import { PermissionEnum } from "../../server/schemas/common.schema";
 
-export const createPlanSchema = z.object({
-  planName: z.string().min(1, "企画名を入力してください"),
+export const createPlanFormSchema = z.object({
+  title: z.string().min(1, "企画名を入力してください"),
   year: z.string().min(1, "対象年を選択してください"),
   month: z.string().min(1, "対象月を選択してください"),
   participants: z.array(
@@ -9,8 +10,9 @@ export const createPlanSchema = z.object({
       userId: z.string(),
       lastName: z.string(),
       firstName: z.string(),
+      permission: z.enum(PermissionEnum),
     })
   ),
 });
 
-export type CreatePlanFormDataType = z.infer<typeof createPlanSchema>;
+export type CreatePlanFormData = z.infer<typeof createPlanFormSchema>;
