+++ src/features/plan/base/create/widgets/form/index.tsx
@@ -1,7 +1,7 @@
 import { FieldWrapper, Select } from "@/components/form";
 import { TextField } from "@/components/form";
 import { useCreatePlan } from "@/features/plan/base/create";
-import { getYearOptions } from "../createPlanutils";
+import { getYearOptions } from "../../utils";
 import AddParticipants from "./addParticipants";
 
 export default function CreatePlanForm() {
